+++ Packages/webxr-interactions/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Check for alwaysUseGrip in ControllerInteraction.
 
 ## [0.6.0] - 2020-01-23
 ### Added