+++ Packages/webxr/CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support for `targetRaySpace` for controllers poses.
+
+### Changed
+- Use `targetRaySpace` as the default controller pose instead of `gripSpace`.
 
 ## [0.6.0] - 2020-01-23
 ### Added